sink(ticdc): move codec into mq
