+++ cdc/sink/mq/codec/avro_test_data.go
