+++ cdc/sink/mq/codec/canal.go
