+++ cdc/sink/mq/codec/canal_flat.go
