+++ cdc/sink/mq/codec/canal_flat_test.go
