+++ cdc/sink/mq/codec/canal_test.go
