+++ cdc/sink/mq/codec/codec_test.go
@@ -21,7 +21,7 @@ import (
 
 	"github.com/pingcap/tidb/parser/mysql"
 	"github.com/pingcap/tiflow/cdc/model"
-	"github.com/pingcap/tiflow/cdc/sink/codec/craft"
+	"github.com/pingcap/tiflow/cdc/sink/mq/codec/craft"
 	"github.com/pingcap/tiflow/proto/benchmark"
 	"github.com/stretchr/testify/require"
 )
