+++ cdc/sink/mq/codec/config.go
