+++ cdc/sink/mq/codec/config_test.go
