+++ cdc/sink/mq/codec/craft/buffer.go
