+++ cdc/sink/mq/codec/craft/codec_test.go
