+++ cdc/sink/mq/codec/craft/decoder.go
