+++ cdc/sink/mq/codec/craft/encoder.go
