+++ cdc/sink/mq/codec/craft/model.go
