+++ cdc/sink/mq/codec/craft_test.go
