+++ cdc/sink/mq/codec/interface.go
