+++ cdc/sink/mq/codec/interface_test.go
