+++ cdc/sink/mq/codec/java.go
