+++ cdc/sink/mq/codec/json.go
