+++ cdc/sink/mq/codec/json_test.go
